Shorten M circuit implementation (#673)
Summary:

In the future, the `nexani.rs` will be replaced by `nexani` crates.
This PR creates a place holder for future usage.

Use circuit gadgets to shorten M circuit implementation

Test Plan:


Summary:

Avoid many repetition in M Chips.

Test Plan:
