+++ prover/src/chips/instructions/m/div_rem.rs
@@ -9,264 +9,14 @@ use crate::{
     traits::MachineChip,
 };
 
-use super::{mul::mul_limb, mulhu::mulh_limb};
-
-pub struct AbsResult {
-    pub abs_limbs: [u8; 4],
-    pub carry: [bool; 2],
-    pub sgn: bool,
-}
-
-/// Compute the absolute value of a 32-bit integer represented as 4 8-bit limbs
-///
-/// This function implements absolute value computation using limb-by-limb operations:
-/// 1. For negative numbers: We negate each limb (two's complement) and add 1
-/// 2. For non-negative numbers: We return the original value
-///
-/// The two's complement negation is done by:
-/// - Inverting each bit (complementing)
-/// - Adding 1 to the result
-///
-/// Returns the absolute value result as limbs and carry flags
-pub fn abs_limb(n: u32) -> AbsResult {
-    //--------------------------------------------------------------
-    // STEP 1: Determine the sign of input and prepare limbs
-    //--------------------------------------------------------------
-    // Extract the sign bit (1 if negative, 0 if positive)
-    let sgn_n = (n >> 31) & 1;
-    assert!(sgn_n < 2, "Sign bit must be 0 or 1");
-
-    // Convert input to individual bytes (limbs)
-    let n_limbs = n.to_le_bytes().map(|x| x as u32);
-    let mut limbs = n.to_le_bytes();
-
-    //--------------------------------------------------------------
-    // STEP 2: Negate using two's complement (for negative numbers)
-    //--------------------------------------------------------------
-    // First complement: invert all bits
-    for l in &mut limbs {
-        *l = u8::MAX - *l;
-    }
-
-    // Second complement: add 1 and propagate carry
-    let mut carry = [false; 4];
-
-    // Add 1 to the least significant limb and propagate carry
-    (limbs[0], carry[0]) = limbs[0].overflowing_add(1);
-    (limbs[1], carry[1]) = limbs[1].overflowing_add(carry[0] as u8);
-    (limbs[2], carry[2]) = limbs[2].overflowing_add(carry[1] as u8);
-    (limbs[3], carry[3]) = limbs[3].overflowing_add(carry[2] as u8);
-
-    //--------------------------------------------------------------
-    // STEP 3: Verify correctness using mathematical constraints
-    //--------------------------------------------------------------
-    // Convert boolean carries to u32 and limbs to u32 for verification
-    let carry_u32 = carry.map(|x| x as u32);
-    let limbs_u32 = limbs.map(|x| x as u32);
-
-    // Verify lower 16 bits correctness
-    let unsigned_lower: u32 = limbs_u32[0] + (limbs_u32[1] << 8);
-    let signed_lower: u32 = (1u32 << 16)
-        .wrapping_sub(n_limbs[0])
-        .wrapping_sub(n_limbs[1] << 8)
-        .wrapping_sub((carry_u32[1]) << 16);
-
-    // This equation verifies correct two's complement calculation for lower 16 bits
-    assert_eq!(
-        (1 - sgn_n) * unsigned_lower + sgn_n * signed_lower,
-        unsigned_lower,
-        "Lower 16 bits verification failed"
-    );
-
-    // Verify upper 16 bits correctness
-    let unsigned_upper: u32 = limbs_u32[2] + (limbs_u32[3] << 8);
-    let signed_upper: u32 = (1u32 << 16)
-        .wrapping_sub(1)
-        .wrapping_sub(n_limbs[2])
-        .wrapping_sub(n_limbs[3] << 8)
-        .wrapping_add(carry_u32[1])
-        .wrapping_sub((carry_u32[3]) << 16);
-
-    // This equation verifies correct two's complement calculation for upper 16 bits
-    assert_eq!(
-        (1 - sgn_n) * unsigned_upper + sgn_n * signed_upper,
-        unsigned_upper,
-        "Upper 16 bits verification failed"
-    );
-
-    //--------------------------------------------------------------
-    // STEP 4: Return the absolute value
-    //--------------------------------------------------------------
-    // Early return for non-negative numbers
-    if sgn_n == 0 {
-        AbsResult {
-            sgn: sgn_n == 1,
-            abs_limbs: n.to_le_bytes(),
-            carry: [false, false],
-        }
-    } else {
-        // For negative input, return the computed absolute value
-        AbsResult {
-            sgn: sgn_n == 1,
-            abs_limbs: limbs,
-            carry: [carry[1], carry[3]], // Store only the important carry bits
-        }
-    }
-}
-
-pub struct DivResult {
-    // Remainder: r = a - b * c = a - t
-    pub r: [u8; 4],
-    // Borrows for r = a - t calculation
-    pub r_borrow: [bool; 2],
-    // Check value: u = c - r - 1
-    pub u: [u8; 4],
-    // Borrows for u = c - r - 1 calculation
-    pub u_borrow: [bool; 2],
-}
-
-// This function verifies the relationship a = b * c + r, where 0 <= r < c.
-// It computes t = b * c, r = a - t, and u = c - r - 1.
-// It asserts that the intermediate borrow flags are as expected,
-// specifically that r >= 0 (borrow_1 is false) and u >= 0 (borrow_3 is false).
-// This implies 0 <= r < c.
-//
-// Preconditions:
-// - b * c must fit within a u32 (checked by mul_limb result t.a_h == [0; 4]).
-// - The specific implementation details of mul_limb might impose further constraints
-//   (e.g., related to t.c5, t.p3_prime_prime, etc.), which are asserted here.
-pub fn div_limb(quotient: u32, dividend: u32, divisor: u32) -> DivResult {
-    let dividend_bytes = dividend.to_le_bytes();
-    let divisor_bytes = divisor.to_le_bytes();
-    let quotient_bytes = quotient.to_le_bytes();
-
-    // When divisor is 0, for DIVU, the result is 2^32 - 1
-    // For REMU, the result is the dividend
-    if divisor == 0 {
-        return DivResult {
-            r: [
-                dividend_bytes[0],
-                dividend_bytes[1],
-                dividend_bytes[2],
-                dividend_bytes[3],
-            ],
-            r_borrow: [false, false],
-            // u = c - r - 1 >= 0 is not possible when c = 0, thus we set u to 0
-            u: [0, 0, 0, 0],
-            u_borrow: [false, false],
-        };
-    }
-
-    // Calculate t = quotient * divisor
-    let t = mulh_limb(quotient, divisor);
-    // t.a_l contains the low 32 bits of quotient * divisor
-    let quotient_divisor_bytes = t._a_l;
-
-    // Assert that quotient * divisor fits within 32 bits (high part of multiplication is zero)
-    assert_eq!(
-        t._a_h,
-        [0, 0, 0, 0],
-        "Overflow: quotient * divisor exceeds u32::MAX"
-    );
-
-    // Assertions based on mul_limb internal details/constraints
-    // These might be specific to the circuit implementation relying on mul_limb
-    assert_eq!(t.c5, false);
-    assert_eq!(t.p3_prime_prime[1], 0);
-    assert_eq!(t.c3_prime_prime, false);
-    assert_eq!(t.c3_prime, false);
-    assert_eq!(t.p3_prime[1], 0);
-    assert_eq!(t.carry_l[1], 0);
-
-    // The above assertions can be combined into a single assertion
-    {
-        let quotient_bytes = quotient_bytes.map(|x| x as u32);
-        let divisor_bytes = divisor_bytes.map(|x| x as u32);
-        assert_eq!(
-            (quotient_bytes[2] + quotient_bytes[3]) * (divisor_bytes[2] + divisor_bytes[3])
-                + quotient_bytes[1] * divisor_bytes[3]
-                + quotient_bytes[3] * divisor_bytes[1],
-            0
-        );
-    }
-
-    // Calculate r = a - t = (a - bc) using 16-bit chunks
-    let a_low = u16::from_le_bytes([dividend_bytes[0], dividend_bytes[1]]);
-    let bc_low = u16::from_le_bytes([quotient_divisor_bytes[0], quotient_divisor_bytes[1]]);
-    let (r_low, borrow_0) = a_low.borrowing_sub(bc_low, false);
-
-    let a_high = u16::from_le_bytes([dividend_bytes[2], dividend_bytes[3]]);
-    let bc_high = u16::from_le_bytes([quotient_divisor_bytes[2], quotient_divisor_bytes[3]]);
-    // Propagate borrow from the low part subtraction
-    let (r_high, borrow_1) = a_high.borrowing_sub(bc_high, borrow_0);
-
-    // Combine the 16-bit results back into r_bytes
-    let r_bytes = [
-        r_low.to_le_bytes()[0],
-        r_low.to_le_bytes()[1],
-        r_high.to_le_bytes()[0],
-        r_high.to_le_bytes()[1],
-    ];
-
-    // Verify the subtraction r = a - bc
-    let r_val = u32::from_le_bytes(r_bytes);
-    let bc_val = u32::from_le_bytes(quotient_divisor_bytes);
-    assert_eq!(r_val, dividend.wrapping_sub(bc_val));
-
-    // Verify the low part subtraction: r_low + borrow_0 * 2^16 = bc_low + r_low
-    assert_eq!(
-        a_low as u32 + ((borrow_0 as u32) << 16),
-        bc_low as u32 + r_low as u32,
-        "Low part subtraction check failed"
-    );
-
-    // Verify the high part subtraction: a_high + borrow_1 * 2^16 = bc_high + r_high + borrow_0
-    assert_eq!(
-        a_high as u32 + ((borrow_1 as u32) << 16),
-        bc_high as u32 + r_high as u32 + borrow_0 as u32,
-        "High part subtraction check failed"
-    );
-
-    // Check if r >= 0. This is the core requirement for standard division remainder.
-    // If a >= bc, borrow_1 would be false.
-    assert_eq!(borrow_1, false, "Remainder r is negative (a < b*c)");
-
-    // Calculate u = c - r - 1
-    // This check helps ensure r < c. If c - r - 1 >= 0, then c - r > 0, so c > r.
-    let c_low = u16::from_le_bytes([divisor_bytes[0], divisor_bytes[1]]);
-    // Start with an initial borrow of 1 for the "- 1" part
-    let (u_low, borrow_2) = c_low.borrowing_sub(r_low, true);
-
-    let c_high = u16::from_le_bytes([divisor_bytes[2], divisor_bytes[3]]);
-    // Propagate borrow from the low part subtraction
-    let (u_high, borrow_3) = c_high.borrowing_sub(r_high, borrow_2);
-
-    // Combine the 16-bit results back into u_bytes
-    let u_bytes = [
-        u_low.to_le_bytes()[0],
-        u_low.to_le_bytes()[1],
-        u_high.to_le_bytes()[0],
-        u_high.to_le_bytes()[1],
-    ];
-
-    // Verify the subtraction u = c - r - 1
-    let u_val = u_low as u32 + ((u_high as u32) << 16);
-    assert_eq!(u_val, divisor.wrapping_sub(r_val).wrapping_sub(1));
-
-    // Check if u >= 0 (i.e., c - r - 1 >= 0 => c > r).
-    // If c <= r, borrow_3 must be false.
-    assert_eq!(borrow_3, false, "Check u >= 0 failed (c <= r)");
-
-    // An explicit check that c >= r, which is implied by borrow_3 == false if r is non-negative.
-    assert!(divisor >= r_val, "Constraint c >= r failed");
-
-    DivResult {
-        r: r_bytes, // r = a - t
-        r_borrow: [borrow_0, borrow_1],
-        u: u_bytes, // u = c - r - 1
-        u_borrow: [borrow_2, borrow_3],
-    }
-}
+use super::{
+    gadget::{
+        constraint_gadget_abs32, constraint_gadget_is_overflow, constraint_gadget_is_zero,
+        constraint_gadget_mul_product, constraint_gadget_sign_1_to_1,
+        constraint_gadget_sign_2_to_1,
+    },
+    nexani::{abs_limb, divu_limb, mull_limb, AbsResult},
+};
 
 pub struct DivRemChip;
 
@@ -332,7 +82,6 @@ impl MachineChip for DivRemChip {
                 }
                 // Normal division case
                 _ => {
-                    // TODO: convert this result to absolute value
                     let quotient =
                         i32::from_le_bytes(value_b).wrapping_div(i32::from_le_bytes(value_c));
                     let remainder =
@@ -346,7 +95,7 @@ impl MachineChip for DivRemChip {
                 }
             };
 
-        traces.fill_columns(row_idx, is_c_zero, IsCZero);
+        traces.fill_columns(row_idx, is_c_zero, IsDivideByZero);
         traces.fill_columns(row_idx, is_overflow, IsOverflow);
 
         let abs_value_b = abs_limb(u32::from_le_bytes(value_b));
@@ -370,7 +119,7 @@ impl MachineChip for DivRemChip {
         traces.fill_columns(row_idx, abs_value_c.carry, ValueCAbsBorrow);
 
         // Calculate t = quotient * divisor using mul_limb to get intermediate values
-        let mul_result = mul_limb(
+        let mul_result = mull_limb(
             u32::from_le_bytes(quotient.abs_limbs),
             u32::from_le_bytes(abs_value_c.abs_limbs),
         );
@@ -392,11 +141,12 @@ impl MachineChip for DivRemChip {
         traces.fill_columns(row_idx, mul_result.c3_prime_prime, MulC3PrimePrime);
 
         // Store t = quotient * divisor
-        traces.fill_columns(row_idx, mul_result._a_l, HelperT);
+        traces.fill_columns(row_idx, mul_result.a_l, HelperT);
 
         // Calculate the division results (remainder r, check value u)
-        let divu_result = div_limb(
+        let divu_result = divu_limb(
             u32::from_le_bytes(quotient.abs_limbs),    // quotient
+            u32::from_le_bytes(remainder.abs_limbs),   // remainder
             u32::from_le_bytes(abs_value_b.abs_limbs), // dividend
             u32::from_le_bytes(abs_value_c.abs_limbs), // divisor
         );
@@ -414,13 +164,6 @@ impl MachineChip for DivRemChip {
         traces.fill_columns(row_idx, abs_value_b.sgn, SgnB);
         traces.fill_columns(row_idx, abs_value_c.sgn, SgnC);
 
-        let helper_1 = [value_a[0], value_a[1], value_a[2], value_a[3] & 0x7F];
-        let helper_2 = [value_b[0], value_b[1], value_b[2], value_b[3] & 0x7F];
-        let helper_3 = [value_c[0], value_c[1], value_c[2], value_c[3] & 0x7F];
-
-        traces.fill_columns(row_idx, helper_1, Helper1);
-        traces.fill_columns(row_idx, helper_2, Helper2);
-        traces.fill_columns(row_idx, helper_3, Helper3);
         traces.fill_columns(row_idx, vm_step.get_sgn_result(), SgnA);
         traces.fill_columns(row_idx, vm_step.get_sgn_b(), SgnB);
         traces.fill_columns(row_idx, vm_step.get_sgn_c(), SgnC);
@@ -446,218 +189,127 @@ impl MachineChip for DivRemChip {
         let [sgn_b] = trace_eval!(trace_eval, SgnB);
         let [sgn_c] = trace_eval!(trace_eval, SgnC);
         let value_a = trace_eval!(trace_eval, ValueA);
-        let [is_c_zero] = trace_eval!(trace_eval, IsCZero);
+        let [is_divide_by_zero] = trace_eval!(trace_eval, IsDivideByZero);
         let [is_a_zero] = trace_eval!(trace_eval, IsAZero);
         let [is_overflow] = trace_eval!(trace_eval, IsOverflow);
 
         // Check for is_c_zero
         // (is_div + is_rem) ⋅ ((c_0 + c_1 ⋅ 2^8 + c_2 ⋅ 2^16 + c_3 ⋅ 2^22) ⋅ is_c_zero
-        // + (1 - is_c_zero) ⋅ (c_0 + c_1 ⋅ 2^8 + c_2 ⋅ 2^16 + c_3 ⋅ 2^22)
-        // - (c_0 + c_1 ⋅ 2^8 + c_2 ⋅ 2^16 + c_3 ⋅ 2^22))
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * ((value_c[0].clone()
-                    + value_c[1].clone() * BaseField::from(1 << 8)
-                    + value_c[2].clone() * BaseField::from(1 << 16)
-                    + value_c[3].clone() * BaseField::from(1 << 22))
-                    * is_c_zero.clone()),
+        constraint_gadget_is_zero(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            is_divide_by_zero.clone(),
+            value_c.clone(),
         );
         // Check for is_a_zero
         // (is_div + is_rem) ⋅ ((a_0 + a_1 ⋅ 2^8 + a_2 ⋅ 2^16 + a_3 ⋅ 2^22) ⋅ is_a_zero
-        // + (1 - is_a_zero) ⋅ (a_0 + a_1 ⋅ 2^8 + a_2 ⋅ 2^16 + a_3 ⋅ 2^22)
-        // - (a_0 + a_1 ⋅ 2^8 + a_2 ⋅ 2^16 + a_3 ⋅ 2^22))
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * ((value_a[0].clone()
-                    + value_a[1].clone() * BaseField::from(1 << 8)
-                    + value_a[2].clone() * BaseField::from(1 << 16)
-                    + value_a[3].clone() * BaseField::from(1 << 22))
-                    * is_a_zero.clone()),
+        constraint_gadget_is_zero(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            is_a_zero.clone(),
+            value_a.clone(),
         );
 
         // Check for is_overflow when dividend is i32::MIN and divisor is -1
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * (is_overflow.clone() * E::F::one()
-                    + (E::F::one() - is_overflow.clone())
-                        * (value_b[0].clone()
-                            + value_b[1].clone() * BaseField::from(1 << 8)
-                            + value_b[2].clone() * BaseField::from(1 << 16)
-                            + value_b[3].clone() * BaseField::from(1 << 24))
-                    - (value_b[0].clone()
-                        + value_b[1].clone() * BaseField::from(1 << 8)
-                        + value_b[2].clone() * BaseField::from(1 << 16)
-                        + value_b[3].clone() * BaseField::from(1 << 24))
-                    + is_overflow.clone() * E::F::one()
-                    + (E::F::one() - is_overflow.clone())
-                        * (value_c[0].clone()
-                            + value_c[1].clone() * BaseField::from(1 << 8)
-                            + value_c[2].clone() * BaseField::from(1 << 16)
-                            + value_c[3].clone() * BaseField::from(1 << 24))
-                    - (value_c[0].clone()
-                        + value_c[1].clone() * BaseField::from(1 << 8)
-                        + value_c[2].clone() * BaseField::from(1 << 16)
-                        + value_c[3].clone() * BaseField::from(1 << 24))),
+        constraint_gadget_is_overflow(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            is_overflow.clone(),
+            value_b.clone(),
+            value_c.clone(),
         );
 
         // Assert that the committed absolute value_b is equal to dividend
         // (is_div + is_rem) ⋅
         // [(1 − sgn_b) ⋅ (b_0 + b_1 ⋅ 2^8) + sgn_b ⋅ (2^16 − b_0 - b_1 ⋅ 2^8 - abs_value_b_borrow ⋅ 2^16) - dividend_0 - dividend_1 ⋅ 2^8]
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * ((E::F::one() - sgn_b.clone())
-                    * (value_b[0].clone() + value_b[1].clone() * BaseField::from(1 << 8))
-                    + sgn_b.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked(1 << 16))
-                            - (value_b[0].clone() + value_b[1].clone() * BaseField::from(1 << 8))
-                            - abs_value_b_borrow[0].clone() * BaseField::from(1 << 16))
-                    - abs_dividend_b[0].clone()
-                    - abs_dividend_b[1].clone() * BaseField::from(1 << 8)),
-        );
         // (is_div + is_rem) ⋅
         // [(1 − sgn_b) ⋅ (b_2 + b_3 ⋅ 2^8) + sgn_b ⋅ (2^16 - 1 - b_2 - b_3 ⋅ 2^8 - abs_value_b_borrow_1 ⋅ 2^16 + abs_value_b_borrow_0) - dividend_2 - dividend_3 ⋅ 2^8]
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * ((E::F::one() - sgn_b.clone())
-                    * (value_b[2].clone() + value_b[3].clone() * BaseField::from(1 << 8))
-                    + sgn_b.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked((1 << 16) - 1))
-                            - (value_b[2].clone() + value_b[3].clone() * BaseField::from(1 << 8))
-                            - abs_value_b_borrow[1].clone() * BaseField::from(1 << 16)
-                            + abs_value_b_borrow[0].clone())
-                    - abs_dividend_b[2].clone()
-                    - abs_dividend_b[3].clone() * BaseField::from(1 << 8)),
+        constraint_gadget_abs32(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            sgn_b.clone(),
+            value_b.clone(),
+            abs_dividend_b.clone(),
+            abs_value_b_borrow.clone(),
         );
 
         // Assert that the committed absolute value_c is equal to divisor
         // (is_div + is_rem) ⋅
         // [(1 − sgn_c) ⋅ (c_0 + c_1 ⋅ 2^8) + sgn_c ⋅ (2^16 − c_0 - c_1 ⋅ 2^8 - abs_value_c_borrow ⋅ 2^16) - divisor_0 - divisor_1 ⋅ 2^8]
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * ((E::F::one() - sgn_c.clone())
-                    * (value_c[0].clone() + value_c[1].clone() * BaseField::from(1 << 8))
-                    + sgn_c.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked(1 << 16))
-                            - (value_c[0].clone() + value_c[1].clone() * BaseField::from(1 << 8))
-                            - abs_value_c_borrow[0].clone() * BaseField::from(1 << 16))
-                    - abs_divisor_c[0].clone()
-                    - abs_divisor_c[1].clone() * BaseField::from(1 << 8)),
-        );
         // (is_div + is_rem) ⋅
         // [(1 − sgn_c) ⋅ (c_2 + c_3 ⋅ 2^8) + sgn_c ⋅ (2^16 - 1 - c_2 - c_3 ⋅ 2^8 - abs_value_c_borrow_1 ⋅ 2^16 + abs_value_c_borrow_0) - divisor_2 - divisor_3 ⋅ 2^8]
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * ((E::F::one() - sgn_c.clone())
-                    * (value_c[2].clone() + value_c[3].clone() * BaseField::from(1 << 8))
-                    + sgn_c.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked((1 << 16) - 1))
-                            - (value_c[2].clone() + value_c[3].clone() * BaseField::from(1 << 8))
-                            - abs_value_c_borrow[1].clone() * BaseField::from(1 << 16)
-                            + abs_value_c_borrow[0].clone())
-                    - abs_divisor_c[2].clone()
-                    - abs_divisor_c[3].clone() * BaseField::from(1 << 8)),
+        constraint_gadget_abs32(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            sgn_c.clone(),
+            value_c.clone(),
+            abs_divisor_c.clone(),
+            abs_value_c_borrow.clone(),
         );
-
         let quotient = trace_eval!(trace_eval, Quotient);
         let abs_value_a_borrow = trace_eval!(trace_eval, ValueAAbsBorrow);
         let [sgn_a] = trace_eval!(trace_eval, SgnA);
 
         // For REM, the sign of remainder is the same as the sign of dividend, except when overflow occurs
-        eval.add_constraint(
-            is_rem.clone()
-                * (E::F::one() - is_c_zero.clone() - is_overflow.clone())
-                * (E::F::one() - is_a_zero.clone())
-                * (sgn_a.clone() - sgn_b.clone()),
+        constraint_gadget_sign_1_to_1(
+            eval,
+            is_rem.clone() * (E::F::one() - is_divide_by_zero.clone() - is_overflow.clone()),
+            sgn_a.clone(),
+            is_a_zero.clone(),
+            sgn_b.clone(),
         );
 
         // For DIV, the sign of quotient is sign_b xor sign_c, except when valueC is zero and overflow occurs
-        eval.add_constraint(
-            (is_div.clone())
-                * (E::F::one() - is_c_zero.clone() - is_overflow.clone())
-                * (E::F::one() - is_a_zero.clone())
-                * (sgn_a.clone()
-                    - (sgn_b.clone() + sgn_c.clone()
-                        - sgn_b.clone() * sgn_c.clone() * BaseField::from(2))),
+        constraint_gadget_sign_2_to_1(
+            eval,
+            is_div.clone() * (E::F::one() - is_divide_by_zero.clone() - is_overflow.clone()),
+            sgn_a.clone(),
+            is_a_zero.clone(),
+            [sgn_b.clone(), sgn_c.clone()],
         );
-        // Assert that the committed absolute value_a is equal to quotient
+
+        // Assert that the committed absolute value_a is equal to quotient except when valueC is zero or overflow occurs
         // is_div ⋅
         // [(1 − sgn_a) ⋅ (a_0 + a_1 ⋅ 2^8) + sgn_a ⋅ (2^16 − a_0 - a_1 ⋅ 2^8 - abs_value_a_borrow ⋅ 2^16) - quotient_0 - quotient_1 ⋅ 2^8]
-        eval.add_constraint(
-            is_div.clone()
-                * (E::F::one() - is_c_zero.clone())
-                * (E::F::one() - is_overflow.clone())
-                * ((E::F::one() - sgn_a.clone())
-                    * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                    + sgn_a.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked(1 << 16))
-                            - (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                            - abs_value_a_borrow[0].clone() * BaseField::from(1 << 16))
-                    - quotient[0].clone()
-                    - quotient[1].clone() * BaseField::from(1 << 8)),
-        );
-        // [(1 − sgn_a) ⋅ (a_2 + a_3 ⋅ 2^8) + sgn_a ⋅ (2^16 - 1 - a_2 - a_3 ⋅ 2^8 - abs_value_a_borrow_1 ⋅ 2^16 + abs_value_a_borrow_0) - quotient_2 - quotient_3 ⋅ 2^8]
-        eval.add_constraint(
-            is_div.clone()
-                * (E::F::one() - is_c_zero.clone())
-                * (E::F::one() - is_overflow.clone())
-                * ((E::F::one() - sgn_a.clone())
-                    * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                    + sgn_a.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked((1 << 16) - 1))
-                            - (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                            - abs_value_a_borrow[1].clone() * BaseField::from(1 << 16)
-                            + abs_value_a_borrow[0].clone())
-                    - quotient[2].clone()
-                    - quotient[3].clone() * BaseField::from(1 << 8)),
+        constraint_gadget_abs32(
+            eval,
+            is_div.clone() * (E::F::one() - is_divide_by_zero.clone() - is_overflow.clone()),
+            sgn_a.clone(),
+            value_a.clone(),
+            quotient.clone(),
+            abs_value_a_borrow.clone(),
         );
 
         let remainder = trace_eval!(trace_eval, Remainder);
 
         // Assert that the committed absolute value_a is equal to remainder
         // is_rem ⋅ [(1 − sgn_a) ⋅ (a_0 + a_1 ⋅ 2^8) + sgn_a ⋅ (2^16 - a_0 - a_1 ⋅ 2^8 - abs_value_a_borrow_0 ⋅ 2^16) - remainder_0 - remainder_1 ⋅ 2^8]
-        eval.add_constraint(
-            is_rem.clone()
-                * (E::F::one() - is_c_zero.clone())
-                * ((E::F::one() - sgn_a.clone())
-                    * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                    + sgn_a.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked(1 << 16))
-                            - (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                            - abs_value_a_borrow[0].clone() * BaseField::from(1 << 16))
-                    - remainder[0].clone()
-                    - remainder[1].clone() * BaseField::from(1 << 8)),
-        );
         // is_rem ⋅ [(1 − sgn_a) ⋅ (a_2 + a_3 ⋅ 2^8) + sgn_a ⋅ (2^16 - 1 - a_2 - a_3 ⋅ 2^8 - abs_value_a_borrow_1 ⋅ 2^16 + abs_value_a_borrow_0) - remainder_2 - remainder_3 ⋅ 2^8]
-        eval.add_constraint(
-            is_rem.clone()
-                * (E::F::one() - is_c_zero.clone())
-                * ((E::F::one() - sgn_a.clone())
-                    * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                    + sgn_a.clone()
-                        * (E::F::from(BaseField::from_u32_unchecked((1 << 16) - 1))
-                            - (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                            - abs_value_a_borrow[1].clone() * BaseField::from(1 << 16)
-                            + abs_value_a_borrow[0].clone())
-                    - remainder[2].clone()
-                    - remainder[3].clone() * BaseField::from(1 << 8)),
+        constraint_gadget_abs32(
+            eval,
+            is_rem.clone() * (E::F::one() - is_divide_by_zero.clone()),
+            sgn_a.clone(),
+            value_a.clone(),
+            remainder.clone(),
+            abs_value_a_borrow.clone(),
         );
 
         // Handle DIV exception:If C is zero, then the result of the division is `-1`
         // When overflow occurs, the result of the division is `-2^31`
         eval.add_constraint(
             is_div.clone()
-                * ((E::F::one() - is_c_zero.clone() - is_overflow.clone())
+                * ((E::F::one() - is_divide_by_zero.clone() - is_overflow.clone())
                     * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone() * BaseField::from(0xFFFF)
+                    + is_divide_by_zero.clone() * BaseField::from(0xFFFF)
                     + is_overflow.clone() * BaseField::from(0)
                     - value_a[0].clone()
                     - value_a[1].clone() * BaseField::from(1 << 8)),
         );
         eval.add_constraint(
             is_div.clone()
-                * ((E::F::one() - is_c_zero.clone() - is_overflow.clone())
+                * ((E::F::one() - is_divide_by_zero.clone() - is_overflow.clone())
                     * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone() * BaseField::from(0xFFFF)
+                    + is_divide_by_zero.clone() * BaseField::from(0xFFFF)
                     + is_overflow.clone() * BaseField::from(0x8000)
                     - value_a[2].clone()
                     - value_a[3].clone() * BaseField::from(1 << 8)),
@@ -666,9 +318,9 @@ impl MachineChip for DivRemChip {
         // When overflow occurs, the result of the remainder is 0
         eval.add_constraint(
             is_rem.clone()
-                * ((E::F::one() - is_c_zero.clone() - is_overflow.clone())
+                * ((E::F::one() - is_divide_by_zero.clone() - is_overflow.clone())
                     * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone()
+                    + is_divide_by_zero.clone()
                         * (value_b[0].clone() + value_b[1].clone() * BaseField::from(1 << 8))
                     + is_overflow.clone() * BaseField::from(0)
                     - value_a[0].clone()
@@ -676,9 +328,9 @@ impl MachineChip for DivRemChip {
         );
         eval.add_constraint(
             is_rem.clone()
-                * ((E::F::one() - is_c_zero.clone() - is_overflow.clone())
+                * ((E::F::one() - is_divide_by_zero.clone() - is_overflow.clone())
                     * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone()
+                    + is_divide_by_zero.clone()
                         * (value_b[2].clone() + value_b[3].clone() * BaseField::from(1 << 8))
                     + is_overflow.clone() * BaseField::from(0)
                     - value_a[2].clone()
@@ -715,41 +367,47 @@ impl MachineChip for DivRemChip {
 
         // (is_mul + is_mulh + is_mulhu + is_mulhsu + is_div + is_divu + is_rem + is_remu) ⋅
         // [𝑃 ′3 + 𝑐′3 ⋅ 2^16 − (|𝑏|0 + |𝑏|3) ⋅ (|𝑐|0 + |𝑐|3) + 𝑧0 + 𝑧3]
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * (p3_prime[0].clone()
-                    + p3_prime[1].clone() * BaseField::from(1 << 8)
-                    + c3_prime.clone() * BaseField::from(1 << 16)
-                    - (quotient[0].clone() + quotient[3].clone())
-                        * (abs_divisor_c[0].clone() + abs_divisor_c[3].clone())
-                    + z_0.clone()
-                    + z_3.clone()),
+        constraint_gadget_mul_product(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            p3_prime.clone(),
+            c3_prime.clone(),
+            quotient[0].clone(),
+            quotient[3].clone(),
+            abs_divisor_c[0].clone(),
+            abs_divisor_c[3].clone(),
+            z_0.clone(),
+            z_3.clone(),
         );
 
         // (is_mul + is_mulh + is_mulhu + is_mulhsu + is_div + is_divu + is_rem + is_remu) ⋅
         // [𝑃 ″3 + 𝑐″3 ⋅ 2^16 − (|𝑏|1 + |𝑏|2) ⋅ (|𝑐|1 + |𝑐|2) + 𝑧1 + 𝑧2]
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * (p3_prime_prime[0].clone()
-                    + p3_prime_prime[1].clone() * BaseField::from(1 << 8)
-                    + c3_prime_prime.clone() * BaseField::from(1 << 16)
-                    - (quotient[1].clone() + quotient[2].clone())
-                        * (abs_divisor_c[1].clone() + abs_divisor_c[2].clone())
-                    + z_1.clone()
-                    + z_2.clone()),
+        constraint_gadget_mul_product(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            p3_prime_prime.clone(),
+            c3_prime_prime.clone(),
+            quotient[1].clone(),
+            quotient[2].clone(),
+            abs_divisor_c[1].clone(),
+            abs_divisor_c[2].clone(),
+            z_1.clone(),
+            z_2.clone(),
         );
 
         // (is_mul + is_div + is_divu + is_rem + is_remu) ⋅
         // [𝑃 1 + 𝑐1 ⋅ 2^16 − (|𝑏|0 + |𝑏|1) ⋅ (|𝑐|0 + |𝑐|1) + 𝑧0 + 𝑧1]
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * (p1[0].clone()
-                    + p1[1].clone() * BaseField::from(1 << 8)
-                    + c1.clone() * BaseField::from(1 << 16)
-                    - (quotient[0].clone() + quotient[1].clone())
-                        * (abs_divisor_c[0].clone() + abs_divisor_c[1].clone())
-                    + z_0.clone()
-                    + z_1.clone()),
+        constraint_gadget_mul_product(
+            eval,
+            is_div.clone() + is_rem.clone(),
+            p1.clone(),
+            c1.clone(),
+            quotient[0].clone(),
+            quotient[1].clone(),
+            abs_divisor_c[0].clone(),
+            abs_divisor_c[1].clone(),
+            z_0.clone(),
+            z_1.clone(),
         );
 
         let [mul_carry_0] = trace_eval!(trace_eval, MulCarry0);
@@ -825,7 +483,7 @@ impl MachineChip for DivRemChip {
         let helper_u_borrow = trace_eval!(trace_eval, HelperUBorrow); // borrow for c - r - 1
         eval.add_constraint(
             (is_div.clone() + is_rem.clone())
-                * (E::F::one() - is_c_zero.clone())
+                * (E::F::one() - is_divide_by_zero.clone())
                 * ((abs_divisor_c[0].clone() + abs_divisor_c[1].clone() * BaseField::from(1 << 8)) // c_low
                    + helper_u_borrow[0].clone() * BaseField::from(1 << 16) // borrow2 * 2^16
                    - remainder[0].clone() // r_low
@@ -838,7 +496,7 @@ impl MachineChip for DivRemChip {
         // High part: c_high + borrow3 * 2^16 = r_high + u_high + borrow2
         eval.add_constraint(
             (is_div.clone() + is_rem.clone())
-                * (E::F::one() - is_c_zero.clone())
+                * (E::F::one() - is_divide_by_zero.clone())
                 * ((abs_divisor_c[2].clone() + abs_divisor_c[3].clone() * BaseField::from(1 << 8)) // c_high
                     + helper_u_borrow[1].clone() * BaseField::from(1 << 16) // borrow3 * 2^16
                     - remainder[2].clone() // r_high
@@ -850,28 +508,6 @@ impl MachineChip for DivRemChip {
 
         // Assert check value non-negative: u >= 0 (borrow3 must be 0)
         eval.add_constraint((is_div.clone() + is_rem.clone()) * helper_u_borrow[1].clone());
-
-        let helper1_val = trace_eval!(trace_eval, Helper1);
-        let helper2_val = trace_eval!(trace_eval, Helper2);
-        let helper3_val = trace_eval!(trace_eval, Helper3);
-
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * (sgn_a.clone() * BaseField::from(1 << 7) + helper1_val[3].clone()
-                    - value_a[3].clone()),
-        );
-
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * (sgn_b.clone() * BaseField::from(1 << 7) + helper2_val[3].clone()
-                    - value_b[3].clone()),
-        );
-
-        eval.add_constraint(
-            (is_div.clone() + is_rem.clone())
-                * (sgn_c.clone() * BaseField::from(1 << 7) + helper3_val[3].clone()
-                    - value_c[3].clone()),
-        );
     }
 }
