+++ prover/src/chips/instructions/m/divu_remu.rs
@@ -9,163 +9,12 @@ use crate::{
     traits::MachineChip,
 };
 
-use super::{mul::mul_limb, mulhu::mulh_limb};
-
-pub struct DivResult {
-    // Remainder: r = a - b * c = a - t
-    pub r: [u8; 4],
-    // Borrows for r = a - t calculation
-    pub r_borrow: [bool; 2],
-    // Check value: u = c - r - 1
-    pub u: [u8; 4],
-    // Borrows for u = c - r - 1 calculation
-    pub u_borrow: [bool; 2],
-}
-
-// This function verifies the relationship a = b * c + r, where 0 <= r < c.
-// It computes t = b * c, r = a - t, and u = c - r - 1.
-// It asserts that the intermediate borrow flags are as expected,
-// specifically that r >= 0 (borrow_1 is false) and u >= 0 (borrow_3 is false).
-// This implies 0 <= r < c.
-//
-// Preconditions:
-// - b * c must fit within a u32 (checked by mul_limb result t.a_h == [0; 4]).
-// - The specific implementation details of mul_limb might impose further constraints
-//   (e.g., related to t.c5, t.p3_prime_prime, etc.), which are asserted here.
-pub fn divu_limb(quotient: u32, remainder: u32, dividend: u32, divisor: u32) -> DivResult {
-    let dividend_bytes = dividend.to_le_bytes();
-    let divisor_bytes = divisor.to_le_bytes();
-    let quotient_bytes = quotient.to_le_bytes();
-    let remainder_bytes = remainder.to_le_bytes();
-
-    // When divisor is 0, for DIVU, the result is 2^32 - 1
-    // For REMU, the result is the dividend
-    if divisor == 0 {
-        return DivResult {
-            r: [
-                dividend_bytes[0],
-                dividend_bytes[1],
-                dividend_bytes[2],
-                dividend_bytes[3],
-            ],
-            r_borrow: [false, false],
-            // u = c - r - 1 >= 0 is not possible when c = 0, thus we set u to 0
-            u: [0, 0, 0, 0],
-            u_borrow: [false, false],
-        };
-    }
-
-    // Calculate t = quotient * divisor
-    let t = mulh_limb(quotient, divisor);
-    // t.a_l contains the low 32 bits of quotient * divisor
-    let quotient_divisor_bytes = t._a_l;
-
-    // Assert that quotient * divisor fits within 32 bits (high part of multiplication is zero)
-    assert_eq!(
-        t._a_h,
-        [0, 0, 0, 0],
-        "Overflow: quotient * divisor exceeds u32::MAX"
-    );
-
-    // Assertions based on mul_limb internal details/constraints
-    // These might be specific to the circuit implementation relying on mul_limb
-    assert_eq!(t.c5, false);
-    assert_eq!(t.p3_prime_prime[1], 0);
-    assert_eq!(t.c3_prime_prime, false);
-    assert_eq!(t.c3_prime, false);
-    assert_eq!(t.p3_prime[1], 0);
-    assert_eq!(t.carry_l[1], 0);
-
-    // The above assertions can be combined into a single assertion
-    {
-        let quotient_bytes = quotient_bytes.map(|x| x as u32);
-        let divisor_bytes = divisor_bytes.map(|x| x as u32);
-        assert_eq!(
-            (quotient_bytes[2] + quotient_bytes[3]) * (divisor_bytes[2] + divisor_bytes[3])
-                + quotient_bytes[1] * divisor_bytes[3]
-                + quotient_bytes[3] * divisor_bytes[1],
-            0
-        );
-    }
-
-    // Calculate r = a - t = (a - bc) using 16-bit chunks
-    let a_low = u16::from_le_bytes([dividend_bytes[0], dividend_bytes[1]]);
-    let bc_low = u16::from_le_bytes([quotient_divisor_bytes[0], quotient_divisor_bytes[1]]);
-    let (r_low, borrow_0) = a_low.borrowing_sub(bc_low, false);
-
-    let a_high = u16::from_le_bytes([dividend_bytes[2], dividend_bytes[3]]);
-    let bc_high = u16::from_le_bytes([quotient_divisor_bytes[2], quotient_divisor_bytes[3]]);
-    // Propagate borrow from the low part subtraction
-    let (r_high, borrow_1) = a_high.borrowing_sub(bc_high, borrow_0);
-
-    // Combine the 16-bit results back into r_bytes
-    let r_bytes = [
-        r_low.to_le_bytes()[0],
-        r_low.to_le_bytes()[1],
-        r_high.to_le_bytes()[0],
-        r_high.to_le_bytes()[1],
-    ];
-
-    assert_eq!(r_bytes, remainder_bytes, "Remainder is incorrect");
-
-    // Verify the subtraction r = a - bc
-    let r_val = u32::from_le_bytes(r_bytes);
-    let bc_val = u32::from_le_bytes(quotient_divisor_bytes);
-    assert_eq!(r_val, dividend.wrapping_sub(bc_val));
-
-    // Verify the low part subtraction: r_low + borrow_0 * 2^16 = bc_low + r_low
-    assert_eq!(
-        a_low as u32 + ((borrow_0 as u32) << 16),
-        bc_low as u32 + r_low as u32,
-        "Low part subtraction check failed"
-    );
-
-    // Verify the high part subtraction: a_high + borrow_1 * 2^16 = bc_high + r_high + borrow_0
-    assert_eq!(
-        a_high as u32 + ((borrow_1 as u32) << 16),
-        bc_high as u32 + r_high as u32 + borrow_0 as u32,
-        "High part subtraction check failed"
-    );
-    // Check if r >= 0. This is the core requirement for standard division remainder.
-    // If a >= bc, borrow_1 would be false.
-    assert_eq!(borrow_1, false, "Remainder r is negative (a < b*c)");
-
-    // Calculate u = c - r - 1
-    // This check helps ensure r < c. If c - r - 1 >= 0, then c - r > 0, so c > r.
-    let c_low = u16::from_le_bytes([divisor_bytes[0], divisor_bytes[1]]);
-    // Start with an initial borrow of 1 for the "- 1" part
-    let (u_low, borrow_2) = c_low.borrowing_sub(r_low, true);
-
-    let c_high = u16::from_le_bytes([divisor_bytes[2], divisor_bytes[3]]);
-    // Propagate borrow from the low part subtraction
-    let (u_high, borrow_3) = c_high.borrowing_sub(r_high, borrow_2);
-
-    // Combine the 16-bit results back into u_bytes
-    let u_bytes = [
-        u_low.to_le_bytes()[0],
-        u_low.to_le_bytes()[1],
-        u_high.to_le_bytes()[0],
-        u_high.to_le_bytes()[1],
-    ];
-
-    // Verify the subtraction u = c - r - 1
-    let u_val = u_low as u32 + ((u_high as u32) << 16);
-    assert_eq!(u_val, divisor.wrapping_sub(r_val).wrapping_sub(1));
-
-    // Check if u >= 0 (i.e., c - r - 1 >= 0 => c > r).
-    // If c <= r, borrow_3 must be false.
-    assert_eq!(borrow_3, false, "Check u >= 0 failed (c <= r)");
-
-    // An explicit check that c >= r, which is implied by borrow_3 == false if r is non-negative.
-    assert!(divisor >= r_val, "Constraint c >= r failed");
-
-    DivResult {
-        r: r_bytes, // r = a - t
-        r_borrow: [borrow_0, borrow_1],
-        u: u_bytes, // u = c - r - 1
-        u_borrow: [borrow_2, borrow_3],
-    }
-}
+use super::{
+    gadget::{
+        constraint_gadget_is_equal, constraint_gadget_is_zero, constraint_gadget_mul_product,
+    },
+    nexani::{divu_limb, mull_limb},
+};
 
 pub struct DivuRemuChip;
 
@@ -219,7 +68,7 @@ impl MachineChip for DivuRemuChip {
         traces.fill_columns(row_idx, quotient, Quotient);
 
         // Calculate t = quotient * divisor using mul_limb to get intermediate values
-        let mul_result = mul_limb(u32::from_le_bytes(quotient), u32::from_le_bytes(value_c));
+        let mul_result = mull_limb(u32::from_le_bytes(quotient), u32::from_le_bytes(value_c));
 
         // Fill in the intermediate values from mul_limb(quotient, divisor) into traces
         // These are needed for the multiplication constraints that verify t = quotient * divisor
@@ -237,9 +86,9 @@ impl MachineChip for DivuRemuChip {
         traces.fill_columns(row_idx, mul_result.c3_prime, MulC3Prime);
         traces.fill_columns(row_idx, mul_result.c3_prime_prime, MulC3PrimePrime);
 
-        traces.fill_columns(row_idx, value_c == [0, 0, 0, 0], IsCZero);
+        traces.fill_columns(row_idx, value_c == [0, 0, 0, 0], IsDivideByZero);
         // Store t = quotient * divisor
-        traces.fill_columns(row_idx, mul_result._a_l, HelperT);
+        traces.fill_columns(row_idx, mul_result.a_l, HelperT);
 
         // Calculate the division results (remainder r, check value u)
         let divu_result = divu_limb(
@@ -268,65 +117,54 @@ impl MachineChip for DivuRemuChip {
     ) {
         let [is_divu] = trace_eval!(trace_eval, IsDivu);
         let [is_remu] = trace_eval!(trace_eval, IsRemu);
-        let [is_c_zero] = trace_eval!(trace_eval, IsCZero);
+        let [is_divide_by_zero] = trace_eval!(trace_eval, IsDivideByZero);
         let dividend = trace_eval!(trace_eval, ValueB);
         let divisor_c = trace_eval!(trace_eval, ValueC);
         let value_a = trace_eval!(trace_eval, ValueA);
 
-        // Check for is_c_zero
-        // (is_div + is_rem) ⋅ ((c_0 + c_1 ⋅ 2^8 + c_2 ⋅ 2^16 + c_3 ⋅ 2^22) ⋅ is_c_zero
-        // + (1 - is_c_zero) ⋅ (c_0 + c_1 ⋅ 2^8 + c_2 ⋅ 2^16 + c_3 ⋅ 2^22)
-        // - (c_0 + c_1 ⋅ 2^8 + c_2 ⋅ 2^16 + c_3 ⋅ 2^22))
-        eval.add_constraint(
-            (is_divu.clone() + is_remu.clone())
-                * (is_c_zero.clone()
-                    * (divisor_c[0].clone()
-                        + divisor_c[1].clone() * BaseField::from(1 << 8)
-                        + divisor_c[2].clone() * BaseField::from(1 << 16)
-                        + divisor_c[3].clone() * BaseField::from(1 << 22))
-                    + (E::F::one() - is_c_zero.clone())
-                        * (divisor_c[0].clone()
-                            + divisor_c[1].clone() * BaseField::from(1 << 8)
-                            + divisor_c[2].clone() * BaseField::from(1 << 16)
-                            + divisor_c[3].clone() * BaseField::from(1 << 22))
-                    - (divisor_c[0].clone()
-                        + divisor_c[1].clone() * BaseField::from(1 << 8)
-                        + divisor_c[2].clone() * BaseField::from(1 << 16)
-                        + divisor_c[3].clone() * BaseField::from(1 << 22))),
+        // Check for is_divide_by_zero
+        // (is_div + is_rem) ⋅ ((c_0 + c_1 ⋅ 2^8 + c_2 ⋅ 2^16 + c_3 ⋅ 2^22) ⋅ is_divide_by_zero
+        constraint_gadget_is_zero(
+            eval,
+            is_divu.clone() + is_remu.clone(),
+            is_divide_by_zero.clone(),
+            divisor_c.clone(),
         );
 
-        // Handle DIVU exception:If C is zero, then the result of the division is 2^32 - 1
+        // Handle DIVU exception:If divide by zero is true, then the result of the division is 2^32 - 1
+        // otherwise, the result of the division is the quotient
         eval.add_constraint(
             is_divu.clone()
-                * ((E::F::one() - is_c_zero.clone())
+                * ((E::F::one() - is_divide_by_zero.clone())
                     * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone() * E::F::from(BaseField::from(0xFFFF))
+                    + is_divide_by_zero.clone() * E::F::from(BaseField::from(0xFFFF))
                     - value_a[0].clone()
                     - value_a[1].clone() * BaseField::from(1 << 8)),
         );
         eval.add_constraint(
             is_divu.clone()
-                * ((E::F::one() - is_c_zero.clone())
+                * ((E::F::one() - is_divide_by_zero.clone())
                     * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone() * E::F::from(BaseField::from(0xFFFF))
+                    + is_divide_by_zero.clone() * E::F::from(BaseField::from(0xFFFF))
                     - value_a[2].clone()
                     - value_a[3].clone() * BaseField::from(1 << 8)),
         );
-        // Handle REMU exception:If C is zero, then the result of the remainder is the dividend
+        // Handle REMU exception:If divide by zero is true, then the result of the remainder is the dividend
+        // otherwise, the result of the remainder is the dividend
         eval.add_constraint(
             is_remu.clone()
-                * ((E::F::one() - is_c_zero.clone())
+                * ((E::F::one() - is_divide_by_zero.clone())
                     * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone()
+                    + is_divide_by_zero.clone()
                         * (dividend[0].clone() + dividend[1].clone() * BaseField::from(1 << 8))
                     - value_a[0].clone()
                     - value_a[1].clone() * BaseField::from(1 << 8)),
         );
         eval.add_constraint(
             is_remu.clone()
-                * ((E::F::one() - is_c_zero.clone())
+                * ((E::F::one() - is_divide_by_zero.clone())
                     * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8))
-                    + is_c_zero.clone()
+                    + is_divide_by_zero.clone()
                         * (dividend[2].clone() + dividend[3].clone() * BaseField::from(1 << 8))
                     - value_a[2].clone()
                     - value_a[3].clone() * BaseField::from(1 << 8)),
@@ -335,31 +173,9 @@ impl MachineChip for DivuRemuChip {
         let quotient = trace_eval!(trace_eval, Quotient);
         let remainder = trace_eval!(trace_eval, Remainder);
         // Assert that the committed Quotient is equal to value_a
-        eval.add_constraint(
-            is_divu.clone()
-                * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8)
-                    - quotient[0].clone()
-                    - quotient[1].clone() * BaseField::from(1 << 8)),
-        );
-        eval.add_constraint(
-            is_divu.clone()
-                * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8)
-                    - quotient[2].clone()
-                    - quotient[3].clone() * BaseField::from(1 << 8)),
-        );
+        constraint_gadget_is_equal(eval, is_divu.clone(), value_a.clone(), quotient.clone());
         // Assert that the committed Remainder is equal to value_a
-        eval.add_constraint(
-            is_remu.clone()
-                * (value_a[0].clone() + value_a[1].clone() * BaseField::from(1 << 8)
-                    - remainder[0].clone()
-                    - remainder[1].clone() * BaseField::from(1 << 8)),
-        );
-        eval.add_constraint(
-            is_remu.clone()
-                * (value_a[2].clone() + value_a[3].clone() * BaseField::from(1 << 8)
-                    - remainder[2].clone()
-                    - remainder[3].clone() * BaseField::from(1 << 8)),
-        );
+        constraint_gadget_is_equal(eval, is_remu.clone(), value_a.clone(), remainder.clone());
 
         // Now, we verify the committed quotient and remainder are correct
         // We do this by verifying the following constraints:
@@ -391,41 +207,47 @@ impl MachineChip for DivuRemuChip {
 
         // (is_mul + is_mulh + is_mulhu + is_mulhsu + is_div + is_divu + is_rem + is_remu) ⋅
         // [𝑃 ′3 + 𝑐′3 ⋅ 2^16 − (|𝑏|0 + |𝑏|3) ⋅ (|𝑐|0 + |𝑐|3) + 𝑧0 + 𝑧3]
-        eval.add_constraint(
-            (is_divu.clone() + is_remu.clone())
-                * (p3_prime[0].clone()
-                    + p3_prime[1].clone() * BaseField::from(1 << 8)
-                    + c3_prime.clone() * BaseField::from(1 << 16)
-                    - (quotient[0].clone() + quotient[3].clone())
-                        * (divisor_c[0].clone() + divisor_c[3].clone())
-                    + z_0.clone()
-                    + z_3.clone()),
+        constraint_gadget_mul_product(
+            eval,
+            is_divu.clone() + is_remu.clone(),
+            p3_prime.clone(),
+            c3_prime.clone(),
+            quotient[0].clone(),
+            quotient[3].clone(),
+            divisor_c[0].clone(),
+            divisor_c[3].clone(),
+            z_0.clone(),
+            z_3.clone(),
         );
 
         // (is_mul + is_mulh + is_mulhu + is_mulhsu + is_div + is_divu + is_rem + is_remu) ⋅
         // [𝑃 ″3 + 𝑐″3 ⋅ 2^16 − (|𝑏|1 + |𝑏|2) ⋅ (|𝑐|1 + |𝑐|2) + 𝑧1 + 𝑧2]
-        eval.add_constraint(
-            (is_divu.clone() + is_remu.clone())
-                * (p3_prime_prime[0].clone()
-                    + p3_prime_prime[1].clone() * BaseField::from(1 << 8)
-                    + c3_prime_prime.clone() * BaseField::from(1 << 16)
-                    - (quotient[1].clone() + quotient[2].clone())
-                        * (divisor_c[1].clone() + divisor_c[2].clone())
-                    + z_1.clone()
-                    + z_2.clone()),
+        constraint_gadget_mul_product(
+            eval,
+            is_divu.clone() + is_remu.clone(),
+            p3_prime_prime.clone(),
+            c3_prime_prime.clone(),
+            quotient[1].clone(),
+            quotient[2].clone(),
+            divisor_c[1].clone(),
+            divisor_c[2].clone(),
+            z_1.clone(),
+            z_2.clone(),
         );
 
         // (is_mul + is_div + is_divu + is_rem + is_remu) ⋅
         // [𝑃 1 + 𝑐1 ⋅ 2^16 − (|𝑏|0 + |𝑏|1) ⋅ (|𝑐|0 + |𝑐|1) + 𝑧0 + 𝑧1]
-        eval.add_constraint(
-            (is_divu.clone() + is_remu.clone())
-                * (p1[0].clone()
-                    + p1[1].clone() * BaseField::from(1 << 8)
-                    + c1.clone() * BaseField::from(1 << 16)
-                    - (quotient[0].clone() + quotient[1].clone())
-                        * (divisor_c[0].clone() + divisor_c[1].clone())
-                    + z_0.clone()
-                    + z_1.clone()),
+        constraint_gadget_mul_product(
+            eval,
+            is_divu.clone() + is_remu.clone(),
+            p1.clone(),
+            c1.clone(),
+            quotient[0].clone(),
+            quotient[1].clone(),
+            divisor_c[0].clone(),
+            divisor_c[1].clone(),
+            z_0.clone(),
+            z_1.clone(),
         );
 
         let [mul_carry_0] = trace_eval!(trace_eval, MulCarry0);
@@ -501,7 +323,7 @@ impl MachineChip for DivuRemuChip {
         let helper_u_borrow = trace_eval!(trace_eval, HelperUBorrow); // borrow for c - r - 1
         eval.add_constraint(
             (is_divu.clone() + is_remu.clone())
-                * (E::F::one() - is_c_zero.clone())
+                * (E::F::one() - is_divide_by_zero.clone())
                 * ((divisor_c[0].clone() + divisor_c[1].clone() * BaseField::from(1 << 8)) // c_low
                    + helper_u_borrow[0].clone() * BaseField::from(1 << 16) // borrow2 * 2^16
                    - remainder[0].clone() // r_low
@@ -514,7 +336,7 @@ impl MachineChip for DivuRemuChip {
         // High part: c_high + borrow3 * 2^16 = r_high + u_high + borrow2
         eval.add_constraint(
             (is_divu.clone() + is_remu.clone())
-                * (E::F::one() - is_c_zero.clone())
+                * (E::F::one() - is_divide_by_zero.clone())
                 * ((divisor_c[2].clone() + divisor_c[3].clone() * BaseField::from(1 << 8)) // c_high
                     + helper_u_borrow[1].clone() * BaseField::from(1 << 16) // borrow3 * 2^16
                     - remainder[2].clone() // r_high
