+++ prover/src/chips/instructions/m/mod.rs
@@ -12,3 +12,8 @@ pub use mulh_mulhsu::MulhMulhsuChip;
 
 mod mulhu;
 pub use mulhu::MulhuChip;
+
+// TODO: Move this to nexani crate in the future.
+mod nexani;
+
+mod gadget;
