+++ prover/src/chips/range_check/range_bool.rs
@@ -5,15 +5,16 @@ use num_traits::One;
 use crate::{
     column::Column::{
         self, BorrowFlag, CH1Minus, CH2Minus, CH3Minus, CarryFlag, HelperUBorrow, ImmC, IsAZero,
-        IsAdd, IsAnd, IsAuipc, IsBeq, IsBge, IsBgeu, IsBlt, IsBltu, IsBne, IsCZero, IsDiv, IsDivu,
-        IsEbreak, IsEcall, IsJal, IsJalr, IsLb, IsLbu, IsLh, IsLhu, IsLui, IsLw, IsMul, IsMulh,
-        IsMulhsu, IsMulhu, IsOr, IsOverflow, IsPadding, IsRem, IsRemu, IsSb, IsSh, IsSll, IsSlt,
-        IsSltu, IsSra, IsSrl, IsSub, IsSw, IsSysCycleCount, IsSysDebug, IsSysHalt, IsSysHeapReset,
-        IsSysPrivInput, IsSysStackReset, IsXor, LtFlag, MulC1, MulC3Prime, MulC3PrimePrime, MulC5,
-        MulCarry0, MulCarry1_0, MulCarry1_1, MulCarry2_0, MulCarry2_1, MulCarry3, OpA0, OpB0, OpB4,
-        OpC0, OpC11, OpC12, OpC20, OpC4, PcCarry, ProgCtrCarry, RemAux, RemainderBorrow, SgnA,
-        SgnB, SgnC, ShiftBit1, ShiftBit2, ShiftBit3, ShiftBit4, ShiftBit5, ValueAAbsBorrow,
-        ValueAAbsBorrowHigh, ValueAEffectiveFlag, ValueBAbsBorrow, ValueCAbsBorrow,
+        IsAdd, IsAnd, IsAuipc, IsBeq, IsBge, IsBgeu, IsBlt, IsBltu, IsBne, IsDiv, IsDivideByZero,
+        IsDivu, IsEbreak, IsEcall, IsJal, IsJalr, IsLb, IsLbu, IsLh, IsLhu, IsLui, IsLw, IsMul,
+        IsMulh, IsMulhsu, IsMulhu, IsOr, IsOverflow, IsPadding, IsRem, IsRemu, IsSb, IsSh, IsSll,
+        IsSlt, IsSltu, IsSra, IsSrl, IsSub, IsSw, IsSysCycleCount, IsSysDebug, IsSysHalt,
+        IsSysHeapReset, IsSysPrivInput, IsSysStackReset, IsXor, LtFlag, MulC1, MulC3Prime,
+        MulC3PrimePrime, MulC5, MulCarry0, MulCarry1_0, MulCarry1_1, MulCarry2_0, MulCarry2_1,
+        MulCarry3, OpA0, OpB0, OpB4, OpC0, OpC11, OpC12, OpC20, OpC4, PcCarry, ProgCtrCarry,
+        RemAux, RemainderBorrow, SgnA, SgnB, SgnC, ShiftBit1, ShiftBit2, ShiftBit3, ShiftBit4,
+        ShiftBit5, ValueAAbsBorrow, ValueAAbsBorrowHigh, ValueAEffectiveFlag, ValueBAbsBorrow,
+        ValueCAbsBorrow,
     },
     components::AllLookupElements,
     extensions::ExtensionsConfig,
@@ -115,7 +116,7 @@ const TYPE_R_CHECKED_SINGLE: [Column; 16] = [
     MulC3Prime,
     MulC3PrimePrime,
     MulC5,
-    IsCZero,
+    IsDivideByZero,
     IsOverflow,
     IsAZero,
 ];
